allocation optimization
GC causes a ~2 fps hitch when it happens. Logging the frametime every frame caused a lot of allocations, which caused the GC to have to run every ~13 seconds or so. Now it's running every 2-3 minutes. There's definitely further improvements to be made on the allocation front (I think the string concatenation for the time and (ironically) memory displays are the big offenders now?), but this is significantly better.
Remove the log interface and its implementation in favor of just using println because that was intended for the teensy which didn't work with println, and I'm clearly not using the teensy anytime soon.